tricks for speeding up

+++ utils/clustering/docker_driver.py
@@ -18,6 +18,7 @@ if __name__ == '__main__':
     # then get all input.*.txt files in tc_dir
     input_files = glob.glob(os.path.join(tc_dir, 'input.*.txt'))
     soln_printed = False
+    already_timeout = True
     
     # for each input_file, run os.path.join(tc_dir, soln.py) < input_file, use subprocess and feed in 
     for i, input_file in enumerate(input_files):
@@ -31,7 +32,12 @@ if __name__ == '__main__':
                 print(f'Running {soln_file} {input_file} > {output_file}')
             else: 
                 print(f'Running {soln_file} < {input_file} > {output_file}')
-            
+        
+        if already_timeout:
+            with open(output_file, 'w') as f:
+                f.write("Timeout")
+            continue
+        
         try: 
             if open_ended:
                 p = subprocess.run(['python', soln_file, input_file], stdout=open(output_file, 'w'), stderr=subprocess.PIPE, timeout=timeout)
@@ -58,6 +64,7 @@ if __name__ == '__main__':
                 print(f'Timeout running {soln_file} < {input_file} > {output_file}')
             with open(output_file, 'w') as f:
                 f.write("Timeout")
+            already_timeout = True
                 
         except Exception as e:
             if verbose:
